move BinaryTable class to Memory workspace
